requirements.txt
pyautogui
google-genai
python-dotenv
Pillow
pyscreeze
opencv-python
SpeechRecognition
pyaudio